+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added default type.ps1xml file to project.
+- Added Enums and Classes folders to module source folder. 
 
 ## 1.1.0 - 2021-03-02